Refactor cache
